+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - [feat: upgrade springdoc to 2.8.13](https://github.com/Tencent/spring-cloud-tencent/pull/1738)
 - [refactor:optimize auto configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1739)
 - [refactor:optimize config locate.](https://github.com/Tencent/spring-cloud-tencent/pull/1741)
+- [feat:support async metadata transfer.](https://github.com/Tencent/spring-cloud-tencent/pull/1743)